Adjust formatting for save management section
(and a small change for building section). Note that with markdown numbered lists, you can have the numbers be any positive integer followed by a period to enable automatic iteration. For example, each line in your list could have started with `1.`, and the render will still come out with 1-7 as the result.